Refactor the DocSet to be independent of Dash docs
