+++ docoloco/models/__init__.py
@@ -1,2 +1,2 @@
-from .doc_set import DocSet, Doc, Section
+from .base import Doc, DocSet, Section
 from .search_result_model import SearchResultModel
