+++ docoloco/registry.py
@@ -1,35 +1,8 @@
-from abc import ABC
-from pathlib import Path
 from typing import Dict, List
 
-from .models import DocSet, Doc
-
-
-class DocumentationProvider(ABC):
-    def __init__(self) -> None:
-        super().__init__()
-
-        self.name: str = None
-        self.docs: Dict[str, DocSet] = dict()
-        self.root_path: Path = None
-
-    def load(self) -> None:
-        pass
-
-
-class DashProvider(DocumentationProvider):
-    def __init__(self) -> None:
-        super().__init__()
-        self.name = "Dash"
-        self.root_path = Path(f"{Path.home()}/.local/share/Zeal/Zeal/docsets/")
-
-    def load(self):
-        for doc_path in self.root_path.iterdir():
-            try:
-                doc = DocSet(path=doc_path)
-                self.docs[doc.name] = doc
-            except Exception as e:
-                print(e)
+from .models import Doc, DocSet
+from .providers.base import DocumentationProvider
+from .providers.dash_provider import DashProvider
 
 
 class Registry:
